for #50: Use correct NodeID value when invoking event listeners
The newly added cache entry event listeners receive the node ID of the cluster node on which the event originated. Hazelcast uses different node ID values than Openfire (although confusingly, they're both a byte array that is a UUID). The event listeners should be invoked with Openfire's IDs, not Hazelcast's.